LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/api/origin-from-global.any-expected.txt


PASS Origin.from(globalThis) is a tuple origin.
